+++ frontend/components/collection/layouts/card/CardItem.tsx
@@ -58,21 +58,24 @@ export const CardItem = forwardRef((props: CardItemProps, ref) => {
         transition="border-color 0.2s"
       >
         <HStack w="full" justifyContent="space-between" alignItems="flex-start">
-          <LinkOverlay
-            as={NextLink}
-            href={`/collection/${collection.id}/article/${id}`}
+          <NextLink
             passHref
+            legacyBehavior
+            href={`/?collectionId=${collection.id}&itemId=${id}`}
+            as={`/collection/${collection.id}/article/${id}`}
           >
-            <Heading
-              fontSize={['lg', '2xl']}
-              p={3}
-              fontWeight={Boolean(dateRead) ? 500 : 700}
-              data-test="title"
-              data-test-read={Boolean(dateRead)}
-            >
-              {title}
-            </Heading>
-          </LinkOverlay>
+            <LinkOverlay>
+              <Heading
+                fontSize={['lg', '2xl']}
+                p={3}
+                fontWeight={Boolean(dateRead) ? 500 : 700}
+                data-test="title"
+                data-test-read={Boolean(dateRead)}
+              >
+                {title}
+              </Heading>
+            </LinkOverlay>
+          </NextLink>
 
           {/*<IconButton
             icon={onReadingList ? <BsBookmarkDash /> : <BsBookmarkPlus />}