+++ frontend/features/Panes/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useRefreshActiveCollection';
+export * from './useMarkActiveCollectionAsRead';