+++ frontend/features/Panes/modals/ModalContext.tsx
@@ -2,9 +2,11 @@ import { Dispatch, SetStateAction, useCallback, useState } from 'react';
 import { createContext } from 'use-context-selector';
 import { Collection, ID, noop } from '@shared';
 import { ReactFCC, useToggle } from '@utils';
+import { useDisableHotKeys } from '@features/HotKeys/useDisableHotKeys';
 import { AddModalState } from './AddModal';
+import { useHotkeysContext } from 'react-hotkeys-hook';
 
-type Elements = 'Drawer' | 'AddModal' | 'DeleteModal';
+type Elements = 'Drawer' | 'AddModal' | 'DeleteModal' | 'HotkeysModal';
 
 export type ModalContextData = {
   [P in Elements as `is${P}Open`]: boolean;
@@ -18,21 +20,28 @@ export type ModalContextData = {
   deleteModalState?: ID;
   setDeleteModalState: Dispatch<SetStateAction<ID | undefined>>;
   openDeleteModalWithData: (collectionId: ID) => void;
+
+  // last active scopes
+  hotkeysModalState: string[];
 };
 
 export const ModalContext = createContext<ModalContextData>({
   isDrawerOpen: false,
   isAddModalOpen: false,
   isDeleteModalOpen: false,
+  isHotkeysModalOpen: false,
   closeAddModal: noop,
   closeDrawer: noop,
   closeDeleteModal: noop,
+  closeHotkeysModal: noop,
   toggleAddModal: noop,
   toggleDrawer: noop,
   toggleDeleteModal: noop,
+  toggleHotkeysModal: noop,
   openAddModal: noop,
   openDrawer: noop,
   openDeleteModal: noop,
+  openHotkeysModal: noop,
 
   addModalState: {} as AddModalState,
   setAddModalState: noop,
@@ -40,6 +49,8 @@ export const ModalContext = createContext<ModalContextData>({
 
   setDeleteModalState: noop,
   openDeleteModalWithData: noop,
+
+  hotkeysModalState: [],
 });
 
 export const ModalContextProvider: ReactFCC = ({ children }) => {
@@ -52,6 +63,7 @@ export const ModalContextProvider: ReactFCC = ({ children }) => {
 
   const addModalProps = useAddModal();
   const deleteModalProps = useDeleteModal();
+  const hotkeysModalProps = useHotkeysModal();
 
   return (
     <ModalContext.Provider
@@ -63,6 +75,7 @@ export const ModalContextProvider: ReactFCC = ({ children }) => {
 
         ...addModalProps,
         ...deleteModalProps,
+        ...hotkeysModalProps,
       }}
     >
       {children}
@@ -71,13 +84,24 @@ export const ModalContextProvider: ReactFCC = ({ children }) => {
 };
 
 const useAddModal = () => {
+  const { disableHotkeys, enableHotkeys } = useDisableHotKeys();
   const [addModalState, setAddModalState] = useState<AddModalState>({});
   const {
     isOpen: isAddModalOpen,
-    open: openAddModal,
-    close: closeAddModal,
+    open,
+    close,
     toggle: toggleAddModal,
   } = useToggle();
+
+  const openAddModal = useCallback(() => {
+    open();
+    disableHotkeys();
+  }, [disableHotkeys, open]);
+
+  const closeAddModal = useCallback(() => {
+    close();
+    enableHotkeys();
+  }, [close, enableHotkeys]);
 
   const openAddModalWithData = useCallback(
     (initialData?: Collection) => {
@@ -107,14 +131,25 @@ const useAddModal = () => {
 };
 
 const useDeleteModal = () => {
+  const { disableHotkeys, enableHotkeys } = useDisableHotKeys();
   // state is just ID to remove
   const [deleteModalState, setDeleteModalState] = useState<ID>();
   const {
     isOpen: isDeleteModalOpen,
-    open: openDeleteModal,
-    close: closeDeleteModal,
+    open,
+    close,
     toggle: toggleDeleteModal,
   } = useToggle();
+
+  const openDeleteModal = useCallback(() => {
+    open();
+    disableHotkeys();
+  }, [disableHotkeys, open]);
+
+  const closeDeleteModal = useCallback(() => {
+    close();
+    enableHotkeys();
+  }, [close, enableHotkeys]);
 
   const openDeleteModalWithData = useCallback(
     (id: ID) => {
@@ -132,5 +167,37 @@ const useDeleteModal = () => {
     closeDeleteModal,
     toggleDeleteModal,
     openDeleteModalWithData,
+  };
+};
+
+const useHotkeysModal = () => {
+  const { disableHotkeys, enableHotkeys } = useDisableHotKeys();
+  const { enabledScopes } = useHotkeysContext();
+  const [lastActiveScopes, setLastActiveScopes] = useState(enabledScopes);
+
+  const {
+    isOpen: isHotkeysModalOpen,
+    open,
+    close,
+    toggle: toggleHotkeysModal,
+  } = useToggle();
+
+  const openHotkeysModal = useCallback(() => {
+    setLastActiveScopes(enabledScopes);
+    open();
+    disableHotkeys();
+  }, [disableHotkeys, enabledScopes, open]);
+
+  const closeHotkeysModal = useCallback(() => {
+    close();
+    enableHotkeys();
+  }, [close, enableHotkeys]);
+
+  return {
+    isHotkeysModalOpen,
+    openHotkeysModal,
+    closeHotkeysModal,
+    toggleHotkeysModal,
+    hotkeysModalState: lastActiveScopes,
   };
 };